split vendor form into separate component
